+++ src/proxy/interfaces/i.win-sso.facade.factory.ts
@@ -5,6 +5,7 @@ export interface IWinSsoFacadeFactory {
   create(
     securityPackage: string,
     targetHost: string | undefined,
-    peerCert: PeerCertificate | undefined
+    peerCert: PeerCertificate | undefined,
+    flags: number | undefined
   ): IWinSsoFacade;
 }